% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeROCs.R
\name{makeROCs}
\alias{makeROCs}
\title{Multi-ROC Plotting for TCGA Gene Expression Data}
\usage{
makeROCs(type, target)
}
\arguments{
\item{type}{TCGA-supported acronyms that designate the type of cancer.  makeROCs currently supports ACC, BLCA, KIRC, KIRP, LIHC, and THCA. Instead of specifying a type, “Random” can be used to generate a randomly-chosen type of cancer, as well as a randomly chosen target variable}

\item{target}{the variable to be predicted.  makeROCs currently supports prediction of tumor “stage” (which attempts to distinguish stage I tumors from stage II, III, and IV tumors) and the participant’s “gender”.}
}
\description{
This function generates a multi-Receiver Operating Characteristic (ROC) plot using RNA-seq data from The Cancer Genome Atlas (TCGA) database and a user-specified target variable.  TCGA data is imported using the TCGA2STAT package.  Feature selection is performed, and the remaining variables (in this case, genes) are processed by five different machine learning classifiers: LASSO-Logistic, K-Nearest Neighbor, Random Forest, a Radial-Kernal Support Vector Machine, and a Sigmoid-Kernal Support Vector Machine.  An ROC curve is generated from each classifier, and plotted onto one graph.
}
\details{
Cancer type acronyms:
ACC	  Adrenocortical Carcinoma,
BLCA	  Bladder Urothelial Carcinoma,
KIRC	  Kidney Renal Clear Cell Carcinoma,
KIRP	  Kidney Renal Papillary Cell Carcinoma,
LIHC	  Liver Heptocellular Carcinoma,
THCA	  Thyroid Carcinoma
}
\examples{
makeROCs("THCA", "Stage")
makeROCs("BLCA", "Gender")
makeROCs("Random")

}
\author{
Jacob Blamer, \email{jwilliamblamer@gmail.com}
}
